+++ app/evm_utils.go
@@ -1,0 +1,73 @@
+package app
+
+import (
+	"fmt"
+
+	"github.com/evmos/evmos/v16/precompiles/bech32"
+
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/core/vm"
+	"golang.org/x/exp/maps"
+
+	// staking and distribution precompiles
+	distprecompile "github.com/aura-nw/aura/precompiles/distribution"
+	stakingprecompile "github.com/aura-nw/aura/precompiles/staking"
+
+	authzkeeper "github.com/cosmos/cosmos-sdk/x/authz/keeper"
+	bankkeeper "github.com/cosmos/cosmos-sdk/x/bank/keeper"
+	distributionkeeper "github.com/cosmos/cosmos-sdk/x/distribution/keeper"
+	stakingkeeper "github.com/cosmos/cosmos-sdk/x/staking/keeper"
+	channelkeeper "github.com/cosmos/ibc-go/v7/modules/core/04-channel/keeper"
+	ics20precompile "github.com/evmos/evmos/v16/precompiles/ics20"
+	"github.com/evmos/evmos/v16/precompiles/p256"
+	erc20Keeper "github.com/evmos/evmos/v16/x/erc20/keeper"
+	transferkeeper "github.com/evmos/evmos/v16/x/ibc/transfer/keeper"
+)
+
+func Precompiles(
+	chainID string,
+	stakingKeeper stakingkeeper.Keeper,
+	distributionKeeper distributionkeeper.Keeper,
+	bankKeeper bankkeeper.Keeper,
+	erc20Keeper erc20Keeper.Keeper,
+	authzKeeper authzkeeper.Keeper,
+	transferKeeper transferkeeper.Keeper,
+	channelKeeper channelkeeper.Keeper,
+) map[common.Address]vm.PrecompiledContract {
+	// Clone the mapping from the latest EVM fork.
+	precompiles := maps.Clone(vm.PrecompiledContractsBerlin)
+
+	// secp256r1 precompile as per EIP-7212
+	p256Precompile := &p256.Precompile{}
+
+	bech32Precompile, err := bech32.NewPrecompile(6000)
+	if err != nil {
+		panic(fmt.Errorf("failed to instantiate bech32 precompile: %w", err))
+	}
+
+	stakingPrecompile, err := stakingprecompile.NewPrecompile(stakingKeeper, authzKeeper)
+	if err != nil {
+		panic(fmt.Errorf("failed to instantiate staking precompile: %w", err))
+	}
+
+	distributionPrecompile, err := distprecompile.NewPrecompile(distributionKeeper, stakingKeeper, authzKeeper)
+	if err != nil {
+		panic(fmt.Errorf("failed to instantiate distribution precompile: %w", err))
+	}
+
+	ibcTransferPrecompile, err := ics20precompile.NewPrecompile(transferKeeper, channelKeeper, authzKeeper)
+	if err != nil {
+		panic(fmt.Errorf("failed to instantiate ICS20 precompile: %w", err))
+	}
+
+	// Stateless precompiles
+	precompiles[bech32Precompile.Address()] = bech32Precompile
+	precompiles[p256Precompile.Address()] = p256Precompile
+
+	// Stateful precompiles
+	precompiles[stakingPrecompile.Address()] = stakingPrecompile
+	precompiles[distributionPrecompile.Address()] = distributionPrecompile
+	precompiles[ibcTransferPrecompile.Address()] = ibcTransferPrecompile
+
+	return precompiles
+}