+++ x/evmutil/keeper/params.go
@@ -1,0 +1,18 @@
+package keeper
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	"github.com/aura-nw/aura/x/evmutil/types"
+)
+
+// GetParams returns the total set of evm parameters.
+func (k Keeper) GetParams(ctx sdk.Context) (params types.Params) {
+	k.paramSubspace.GetParamSetIfExists(ctx, &params)
+	return params
+}
+
+// SetParams sets the evm parameters to the param space.
+func (k Keeper) SetParams(ctx sdk.Context, params types.Params) {
+	k.paramSubspace.SetParamSet(ctx, &params)
+}